# Optimise LineString's IsSimple